# swagger.yaml
openapi: 3.0.3
info:
  title: Del-Pick API
  description: API for Del-Pick, a food delivery service application enabling customers to order food, drivers to deliver orders, and store owners to manage their menus and orders.
  version: 1.0.0
  contact:
    name: API Support
    email: support@delpick.id
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://delpick.horas-code.my.id/api/v1
    description: Production Server
  - url: http://localhost:6100/api/v1
    description: Development Server
...
paths:
...
  /driver-requests:
    get:
      tags:
        - Driver Requests
      summary: Retrieve driver requests
      description: Retrieves a paginated list of delivery requests for the authenticated driver, optionally filtered by status.
      operationId: getDriverRequests
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/requestStatusParam'
      responses:
        '200':
          description: Driver requests retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverRequestListResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not a driver
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /driver-requests/{requestId}:
    get:
      tags:
        - Driver Requests
      summary: Retrieve driver request by ID
      description: Retrieves detailed information about a specific delivery request, including order and store details.
      operationId: getDriverRequestById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/driverRequestIdParam'
      responses:
        '200':
          description: Driver request details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                message: Driver request details retrieved successfully
                data:
                  request:
                    id: 1
                    orderId: 1
                    driverId: 1
                    status: pending
                  order:
                    id: 1
                    storeId: 1
                    customerId: 1
                    order_status: approved
                  store:
                    id: 1
                    name: Tasty Bites
                  customer:
                    id: 1
                    name: John Doe
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not the assigned driver
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Driver request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Driver Requests
      summary: Respond to driver request
      description: Accepts or rejects a delivery request. Requires driver privileges.
      operationId: respondToDriverRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/driverRequestIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RespondToRequest'
      responses:
        '200':
          description: Driver request processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                statusCode: 200
                message: Driver request processed successfully
                data:
                  id: 1
                  orderId: 1
                  driverId: 1
                  status: accepted
        '400':
          description: Invalid action
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not the assigned driver
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Driver request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Tracking Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    pageParam:
      name: page
      in: query
      description: Page number for pagination
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    limitParam:
      name: limit
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    driverRequestIdParam:
      name: requestId
      in: path
      description: ID of the driver request
      required: true
      schema:
        type: integer
        minimum: 1
    requestStatusParam:
      name: status
      in: query
      description: Filter by driver request status
      required: false
      schema:
        type: string
        enum: [pending, accepted, rejected, expired]
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: Operation successful
        data:
          type: object
          nullable: true
      required:
        - statusCode
        - message
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: An error occurred
        errors:
          type: string
          nullable: true
          example: Detailed error message
      required:
        - statusCode
        - message

    # Authentication Schemas
...
    DriverRequest:
      type: object
      properties:
        id:
          type: integer
          example: 1
        orderId:
          type: integer
          example: 1
        driverId:
          type: integer
          example: 1
        status:
          type: string
          enum: [pending, accepted, rejected, expired]
          example: pending
        createdAt:
          type: string
          format: date-time
          example: "2025-04-29T09:16:52.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-04-29T09:16:52.000Z"
    DriverRequestListResponse:
      type: object
      properties:
        totalItems:
          type: integer
          example: 10
        totalPages:
          type: integer
          example: 1
        currentPage:
          type: integer
          example: 1
        requests:
          type: array
          items:
            $ref: '#/components/schemas/DriverRequest'
...
    RespondToRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum: [accept, reject]
          example: accept

    # Profile Response Schemas
...